Extract DB into separate file for reuse from different test cases
