+++ Tests/CarthageKitTests/ResolverSpec.swift
@@ -6,94 +6,6 @@ import ReactiveSwift
 import Result
 import Tentacle
 
-private let git1 = Dependency.git(GitURL("https://example.com/repo1"))
-private let git2 = Dependency.git(GitURL("https://example.com/repo2.git"))
-private let github1 = Dependency.gitHub(.dotCom, Repository(owner: "gob", name: "1"))
-private let github2 = Dependency.gitHub(.dotCom, Repository(owner: "gob", name: "2"))
-private let github3 = Dependency.gitHub(.dotCom, Repository(owner: "gob", name: "3"))
-
-// swiftlint:disable no_extension_access_modifier
-private extension PinnedVersion {
-	static let v0_1_0 = PinnedVersion("v0.1.0")
-	static let v1_0_0 = PinnedVersion("v1.0.0")
-	static let v1_1_0 = PinnedVersion("v1.1.0")
-	static let v1_2_0 = PinnedVersion("v1.2.0")
-	static let v2_0_0 = PinnedVersion("v2.0.0")
-	static let v2_0_0_beta_1 = PinnedVersion("v2.0.0-beta.1")
-	static let v2_0_1 = PinnedVersion("v2.0.1")
-	static let v3_0_0_beta_1 = PinnedVersion("v3.0.0-beta.1")
-}
-
-private extension SemanticVersion {
-	static let v0_1_0 = SemanticVersion(major: 0, minor: 1, patch: 0)
-	static let v1_0_0 = SemanticVersion(major: 1, minor: 0, patch: 0)
-	static let v1_1_0 = SemanticVersion(major: 1, minor: 1, patch: 0)
-	static let v1_2_0 = SemanticVersion(major: 1, minor: 2, patch: 0)
-	static let v2_0_0 = SemanticVersion(major: 2, minor: 0, patch: 0)
-	static let v2_0_1 = SemanticVersion(major: 2, minor: 0, patch: 1)
-	static let v3_0_0 = SemanticVersion(major: 3, minor: 0, patch: 0)
-}
-// swiftlint:enable no_extension_access_modifier
-
-// swiftlint:enable identifier_name
-private struct DB {
-	var versions: [Dependency: [PinnedVersion: [Dependency: VersionSpecifier]]]
-	var references: [Dependency: [String: PinnedVersion]] = [:]
-
-	func versions(for dependency: Dependency) -> SignalProducer<PinnedVersion, CarthageError> {
-		if let versions = self.versions[dependency] {
-			return .init(versions.keys)
-		} else {
-			return .init(error: .taggedVersionNotFound(dependency))
-		}
-	}
-
-	func dependencies(for dependency: Dependency, version: PinnedVersion) -> SignalProducer<(Dependency, VersionSpecifier), CarthageError> {
-		if let dependencies = self.versions[dependency]?[version] {
-			return .init(dependencies.map { ($0.0, $0.1) })
-		} else {
-			return .empty
-		}
-	}
-
-	func resolvedGitReference(_ dependency: Dependency, reference: String) -> SignalProducer<PinnedVersion, CarthageError> {
-		if let version = references[dependency]?[reference] {
-			return .init(value: version)
-		} else {
-			return .empty
-		}
-	}
-
-	func resolve(
-		_ resolverType: ResolverProtocol.Type,
-		_ dependencies: [Dependency: VersionSpecifier],
-		resolved: [Dependency: PinnedVersion] = [:],
-		updating: Set<Dependency> = []
-	) -> Result<[Dependency: PinnedVersion], CarthageError> {
-		let resolver = resolverType.init(
-			versionsForDependency: self.versions(for:),
-			dependenciesForDependency: self.dependencies(for:version:),
-			resolvedGitReference: self.resolvedGitReference(_:reference:)
-		)
-		return resolver
-			.resolve(
-				dependencies: dependencies,
-				lastResolved: resolved,
-				dependenciesToUpdate: updating.map { $0.name }
-			)
-			.first()!
-	}
-}
-
-extension DB: ExpressibleByDictionaryLiteral {
-	init(dictionaryLiteral elements: (Dependency, [PinnedVersion: [Dependency: VersionSpecifier]])...) {
-		self.init(versions: [:], references: [:])
-		for (key, value) in elements {
-			versions[key] = value
-		}
-	}
-}
-
 private func ==<A: Equatable, B: Equatable>(lhs: [(A, B)], rhs: [(A, B)]) -> Bool {
 	guard lhs.count == rhs.count else { return false }
 	for (lhs, rhs) in zip(lhs, rhs) {
