+++ Tests/CarthageKitTests/Resources/OutdatedDependencies/Cartfile
@@ -1,6 +1,6 @@
-github "1/1"
-github "2/2"
-github "3/3" "2.0"
-github "4/4" "2.0"
-github "5/5" "development"
-github "6/6" "development"
+github "gob/1"
+github "gob/2"
+github "gob/3" "2.0"
+github "gob/4" "2.0"
+github "gob/5" "development"
+github "gob/6" "development"
