+++ Tests/CarthageKitTests/Resources/OutdatedDependencies/Cartfile.resolved
@@ -1,6 +1,6 @@
-github "1/1" "1.0.0"
-github "2/2" "1.0.0"
-github "3/3" "2.0.0"
-github "4/4" "2.0.0"
-github "5/5" "2ea246ae4573538886ffb946d70d141583443734"
-github "6/6" "2ea246ae4573538886ffb946d70d141583443734"
+github "gob/1" "v1.0.0"
+github "gob/2" "v1.0.0"
+github "gob/3" "v2.0.0"
+github "gob/4" "v2.0.0"
+github "gob/5" "2ea246ae4573538886ffb946d70d141583443734"
+github "gob/6" "2ea246ae4573538886ffb946d70d141583443734"
